requirements.txt
pyyaml
python-dotenv
langchain-anthropic
markdown2